+++ examples/connext_dds/network_capture/03_security/c/CMakeLists.txt
@@ -14,6 +14,6 @@ set(CMAKE_MODULE_PATH
 include(ConnextDdsAddExample)
 
 connextdds_add_example(
-    IDL "example2"
+    IDL "network_capture"
     LANG "C"
 )
