+++ examples/connext_dds/network_capture/03_security/c/network_capture_publisher.c
@@ -15,8 +15,8 @@
  * traffic.
  *
  * This example is a simple secure hello world running network capture for both
- * a publisher participant (example2_publisher.c).and a subscriber participant
- * (example2_subscriber.c).
+ * a publisher participant (network_capture_publisher.c).and a subscriber participant
+ * (network_capture_subscriber.c).
  * It shows the basic flow when working with network capture:
  *   - Enabling before anything else.
  *   - Start capturing traffic (for one or all participants).
@@ -30,102 +30,13 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include "ndds/ndds_c.h"
-#include "example2.h"
-#include "example2Support.h"
+#include "network_capture.h"
+#include "network_captureSupport.h"
 #ifdef RTI_STATIC
 #include "security/security_default.h"
 #endif
 
-void example2Listener_on_requested_deadline_missed(
-        void* listener_data,
-        DDS_DataReader* reader,
-        const struct DDS_RequestedDeadlineMissedStatus *status)
-{
-}
-
-void example2Listener_on_requested_incompatible_qos(
-        void* listener_data,
-        DDS_DataReader* reader,
-        const struct DDS_RequestedIncompatibleQosStatus *status)
-{
-}
-
-void example2Listener_on_sample_rejected(
-        void* listener_data,
-        DDS_DataReader* reader,
-        const struct DDS_SampleRejectedStatus *status)
-{
-}
-
-void example2Listener_on_liveliness_changed(
-        void* listener_data,
-        DDS_DataReader* reader,
-        const struct DDS_LivelinessChangedStatus *status)
-{
-}
-
-void example2Listener_on_sample_lost(
-        void* listener_data,
-        DDS_DataReader* reader,
-        const struct DDS_SampleLostStatus *status)
-{
-}
-
-void example2Listener_on_subscription_matched(
-        void* listener_data,
-        DDS_DataReader* reader,
-        const struct DDS_SubscriptionMatchedStatus *status)
-{
-}
-
-void example2Listener_on_data_available(
-        void* listener_data,
-        DDS_DataReader* reader)
-{
-    example2DataReader *example2_reader = NULL;
-    struct example2Seq data_seq = DDS_SEQUENCE_INITIALIZER;
-    struct DDS_SampleInfoSeq info_seq = DDS_SEQUENCE_INITIALIZER;
-    DDS_ReturnCode_t retcode;
-    int i;
-
-    example2_reader = example2DataReader_narrow(reader);
-    if (example2_reader == NULL) {
-        fprintf(stderr, "DataReader narrow error\n");
-        return;
-    }
-
-    retcode = example2DataReader_take(
-            example2_reader,
-            &data_seq,
-            &info_seq,
-            DDS_LENGTH_UNLIMITED,
-            DDS_ANY_SAMPLE_STATE,
-            DDS_ANY_VIEW_STATE, DDS_ANY_INSTANCE_STATE);
-    if (retcode == DDS_RETCODE_NO_DATA) {
-        return;
-    } else if (retcode != DDS_RETCODE_OK) {
-        fprintf(stderr, "take error %d\n", retcode);
-        return;
-    }
-
-    for (i = 0; i < example2Seq_get_length(&data_seq); ++i) {
-        if (DDS_SampleInfoSeq_get_reference(&info_seq, i)->valid_data) {
-            printf("Received data:\n");
-            example2TypeSupport_print_data(
-                    example2Seq_get_reference(&data_seq, i));
-        }
-    }
-
-    retcode = example2DataReader_return_loan(
-            example2_reader,
-            &data_seq,
-            &info_seq);
-    if (retcode != DDS_RETCODE_OK) {
-        fprintf(stderr, "return loan error %d\n", retcode);
-    }
-}
-
-static int subscriber_shutdown(
+static int publisher_shutdown(
         DDS_DomainParticipant *participant)
 {
     DDS_ReturnCode_t retcode;
@@ -166,18 +77,19 @@ static int subscriber_shutdown(
     return status;
 }
 
-static int subscriber_main(int domainId, int sample_count)
+static int publisher_main(int domainId, int sample_count)
 {
     DDS_DomainParticipant *participant = NULL;
-    DDS_Subscriber *subscriber = NULL;
+    DDS_Publisher *publisher = NULL;
     DDS_Topic *topic = NULL;
-    struct DDS_DataReaderListener reader_listener =
-    DDS_DataReaderListener_INITIALIZER;
-    DDS_DataReader *reader = NULL;
+    DDS_DataWriter *writer = NULL;
+    NetworkCaptureDataWriter *NetworkCapture_writer = NULL;
+    NetworkCapture *instance = NULL;
     DDS_ReturnCode_t retcode;
+    DDS_InstanceHandle_t instance_handle = DDS_HANDLE_NIL;
     const char *type_name = NULL;
     int count = 0;
-    struct DDS_Duration_t poll_period = {1, 0}; /* 1 s */
+    struct DDS_Duration_t send_period = {1, 0}; /* 1 s */
     NDDS_Utility_NetworkCaptureParams_t params =
             NDDS_UTILITY_NETWORK_CAPTURE_PARAMETERS_DEFAULT;
 
@@ -195,16 +107,15 @@ static int subscriber_main(int domainId, int sample_count)
 
     /*
      * Configuration of the capture.
-     *   - Parse encrypted content.
-     *   - Remove all user data (serialized and encrypted) from the capture.
-     *   - Defaults: e.g., capture input and output traffic.
+     *   - Only capture input traffic.
+     *   - Defaults: e.g. Do not parse encrypted content (i.e., show traffic
+     *     with protection, as in wire).
      */
-    params.parse_encrypted_content = DDS_BOOLEAN_TRUE;
-    params.dropped_content =
-            NDDS_UTILITY_NETWORK_CAPTURE_CONTENT_ENCRYPTED_DATA;
-    if (!NDDS_Utility_start_network_capture_w_params("subscriber", &params)) {
+    params.traffic = NDDS_UTILITY_NETWORK_CAPTURE_TRAFFIC_IN;
+    if (!NDDS_Utility_start_network_capture_w_params("publisher", &params)) {
         fprintf(stderr, "Error enabling network monitor for all participants\n");
     }
+
 
     /*
      * The example continues as the usual Connext DDS secure HelloWorld.
@@ -218,9 +129,9 @@ static int subscriber_main(int domainId, int sample_count)
             DDS_TheParticipantFactory,
             &participant_qos,
             "SecurityExampleProfiles",
-            "A");
+            "B");
     if (retcode != DDS_RETCODE_OK) {
-        fprintf(stderr, "Unable to get default participant qos\n");
+        fprintf(stderr, "Unable to get participant qos from profile\n");
         return -1;
     }
     retcode = DDS_PropertyQosPolicyHelper_assert_pointer_property(
@@ -239,85 +150,105 @@ static int subscriber_main(int domainId, int sample_count)
             NULL,
             DDS_STATUS_MASK_NONE);
     DDS_DomainParticipantQos_finalize(&participant_qos);
+
 #else
     participant = DDS_DomainParticipantFactory_create_participant_with_profile(
             DDS_TheParticipantFactory,
             domainId,
             "SecurityExampleProfiles",
-            "A",
+            "B",
             NULL,
             DDS_STATUS_MASK_NONE);
 #endif
     if (participant == NULL) {
         fprintf(stderr, "create_participant error\n");
-        subscriber_shutdown(participant);
+        publisher_shutdown(participant);
         return -1;
     }
 
-    subscriber = DDS_DomainParticipant_create_subscriber(
+    publisher = DDS_DomainParticipant_create_publisher(
             participant,
-            &DDS_SUBSCRIBER_QOS_DEFAULT,
+            &DDS_PUBLISHER_QOS_DEFAULT,
             NULL, /* listener */
             DDS_STATUS_MASK_NONE);
-    if (subscriber == NULL) {
-        fprintf(stderr, "create_subscriber error\n");
-        subscriber_shutdown(participant);
+    if (publisher == NULL) {
+        fprintf(stderr, "create_publisher error\n");
+        publisher_shutdown(participant);
         return -1;
     }
 
-    type_name = example2TypeSupport_get_type_name();
-    retcode = example2TypeSupport_register_type(participant, type_name);
+    type_name = NetworkCaptureTypeSupport_get_type_name();
+    retcode = NetworkCaptureTypeSupport_register_type(
+            participant,
+            type_name);
     if (retcode != DDS_RETCODE_OK) {
         fprintf(stderr, "register_type error %d\n", retcode);
-        subscriber_shutdown(participant);
+        publisher_shutdown(participant);
         return -1;
     }
 
     topic = DDS_DomainParticipant_create_topic(
             participant,
-            "Example example2",
+            "Example NetworkCapture using security",
             type_name,
             &DDS_TOPIC_QOS_DEFAULT,
             NULL, /* listener */
             DDS_STATUS_MASK_NONE);
     if (topic == NULL) {
         fprintf(stderr, "create_topic error\n");
-        subscriber_shutdown(participant);
+        publisher_shutdown(participant);
         return -1;
     }
 
-    reader_listener.on_requested_deadline_missed  =
-    example2Listener_on_requested_deadline_missed;
-    reader_listener.on_requested_incompatible_qos =
-    example2Listener_on_requested_incompatible_qos;
-    reader_listener.on_sample_rejected =
-    example2Listener_on_sample_rejected;
-    reader_listener.on_liveliness_changed =
-    example2Listener_on_liveliness_changed;
-    reader_listener.on_sample_lost =
-    example2Listener_on_sample_lost;
-    reader_listener.on_subscription_matched =
-    example2Listener_on_subscription_matched;
-    reader_listener.on_data_available =
-    example2Listener_on_data_available;
+    writer = DDS_Publisher_create_datawriter(
+            publisher,
+            topic,
+            &DDS_DATAWRITER_QOS_DEFAULT,
+            NULL, /* listener */
+            DDS_STATUS_MASK_NONE);
+    if (writer == NULL) {
+        fprintf(stderr, "create_datawriter error\n");
+        publisher_shutdown(participant);
+        return -1;
+    }
+    NetworkCapture_writer = NetworkCaptureDataWriter_narrow(writer);
+    if (NetworkCapture_writer == NULL) {
+        fprintf(stderr, "DataWriter narrow error\n");
+        publisher_shutdown(participant);
+        return -1;
+    }
 
-    reader = DDS_Subscriber_create_datareader(
-            subscriber,
-            DDS_Topic_as_topicdescription(topic),
-            &DDS_DATAREADER_QOS_DEFAULT,
-            &reader_listener,
-            DDS_STATUS_MASK_ALL);
-    if (reader == NULL) {
-        fprintf(stderr, "create_datareader error\n");
-        subscriber_shutdown(participant);
+    instance = NetworkCaptureTypeSupport_create_data_ex(DDS_BOOLEAN_TRUE);
+    if (instance == NULL) {
+        fprintf(stderr, "NetworkCaptureTypeSupport_create_data error\n");
+        publisher_shutdown(participant);
         return -1;
     }
 
     for (count=0; (sample_count == 0) || (count < sample_count); ++count) {
-        printf("example2 subscriber sleeping for %d sec...\n",
-            poll_period.sec);
 
-        NDDS_Utility_sleep(&poll_period);
+        printf("Writing NetworkCapture Secure, count %d\n", count);
+
+        RTIOsapiUtility_snprintf(
+                instance->msg,
+                128,
+                "Hello World Secure (%d)",
+                count);
+
+        retcode = NetworkCaptureDataWriter_write(
+                NetworkCapture_writer,
+                instance,
+                &instance_handle);
+        if (retcode != DDS_RETCODE_OK) {
+            fprintf(stderr, "write error %d\n", retcode);
+        }
+
+        NDDS_Utility_sleep(&send_period);
+    }
+
+    retcode = NetworkCaptureTypeSupport_delete_data_ex(instance, DDS_BOOLEAN_TRUE);
+    if (retcode != DDS_RETCODE_OK) {
+        fprintf(stderr, "NetworkCaptureTypeSupport_delete_data error %d\n", retcode);
     }
 
     /*
@@ -331,7 +262,7 @@ static int subscriber_main(int domainId, int sample_count)
     /* Finalize the parameters to free allocated memory */
     NDDS_Utility_NetworkCaptureParams_t_finalize(&params);
 
-    return subscriber_shutdown(participant);
+    return publisher_shutdown(participant);
 }
 
 int main(int argc, char *argv[])
@@ -346,6 +277,6 @@ int main(int argc, char *argv[])
         sample_count = atoi(argv[2]);
     }
 
-    return subscriber_main(domainId, sample_count);
+    return publisher_main(domainId, sample_count);
 }
