+++ examples/connext_dds/network_capture/03_security/c/network_capture_subscriber.c
@@ -15,8 +15,8 @@
  * traffic.
  *
  * This example is a simple secure hello world running network capture for both
- * a publisher participant (example2_publisher.c).and a subscriber participant
- * (example2_subscriber.c).
+ * a publisher participant (network_capture_publisher.c).and a subscriber participant
+ * (network_capture_subscriber.c).
  * It shows the basic flow when working with network capture:
  *   - Enabling before anything else.
  *   - Start capturing traffic (for one or all participants).
@@ -30,13 +30,102 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include "ndds/ndds_c.h"
-#include "example2.h"
-#include "example2Support.h"
+#include "network_capture.h"
+#include "network_captureSupport.h"
 #ifdef RTI_STATIC
 #include "security/security_default.h"
 #endif
 
-static int publisher_shutdown(
+void NetworkCaptureListener_on_requested_deadline_missed(
+        void* listener_data,
+        DDS_DataReader* reader,
+        const struct DDS_RequestedDeadlineMissedStatus *status)
+{
+}
+
+void NetworkCaptureListener_on_requested_incompatible_qos(
+        void* listener_data,
+        DDS_DataReader* reader,
+        const struct DDS_RequestedIncompatibleQosStatus *status)
+{
+}
+
+void NetworkCaptureListener_on_sample_rejected(
+        void* listener_data,
+        DDS_DataReader* reader,
+        const struct DDS_SampleRejectedStatus *status)
+{
+}
+
+void NetworkCaptureListener_on_liveliness_changed(
+        void* listener_data,
+        DDS_DataReader* reader,
+        const struct DDS_LivelinessChangedStatus *status)
+{
+}
+
+void NetworkCaptureListener_on_sample_lost(
+        void* listener_data,
+        DDS_DataReader* reader,
+        const struct DDS_SampleLostStatus *status)
+{
+}
+
+void NetworkCaptureListener_on_subscription_matched(
+        void* listener_data,
+        DDS_DataReader* reader,
+        const struct DDS_SubscriptionMatchedStatus *status)
+{
+}
+
+void NetworkCaptureListener_on_data_available(
+        void* listener_data,
+        DDS_DataReader* reader)
+{
+    NetworkCaptureDataReader *NetworkCapture_reader = NULL;
+    struct NetworkCaptureSeq data_seq = DDS_SEQUENCE_INITIALIZER;
+    struct DDS_SampleInfoSeq info_seq = DDS_SEQUENCE_INITIALIZER;
+    DDS_ReturnCode_t retcode;
+    int i;
+
+    NetworkCapture_reader = NetworkCaptureDataReader_narrow(reader);
+    if (NetworkCapture_reader == NULL) {
+        fprintf(stderr, "DataReader narrow error\n");
+        return;
+    }
+
+    retcode = NetworkCaptureDataReader_take(
+            NetworkCapture_reader,
+            &data_seq,
+            &info_seq,
+            DDS_LENGTH_UNLIMITED,
+            DDS_ANY_SAMPLE_STATE,
+            DDS_ANY_VIEW_STATE, DDS_ANY_INSTANCE_STATE);
+    if (retcode == DDS_RETCODE_NO_DATA) {
+        return;
+    } else if (retcode != DDS_RETCODE_OK) {
+        fprintf(stderr, "take error %d\n", retcode);
+        return;
+    }
+
+    for (i = 0; i < NetworkCaptureSeq_get_length(&data_seq); ++i) {
+        if (DDS_SampleInfoSeq_get_reference(&info_seq, i)->valid_data) {
+            printf("Received data:\n");
+            NetworkCaptureTypeSupport_print_data(
+                    NetworkCaptureSeq_get_reference(&data_seq, i));
+        }
+    }
+
+    retcode = NetworkCaptureDataReader_return_loan(
+            NetworkCapture_reader,
+            &data_seq,
+            &info_seq);
+    if (retcode != DDS_RETCODE_OK) {
+        fprintf(stderr, "return loan error %d\n", retcode);
+    }
+}
+
+static int subscriber_shutdown(
         DDS_DomainParticipant *participant)
 {
     DDS_ReturnCode_t retcode;
@@ -77,19 +166,18 @@ static int publisher_shutdown(
     return status;
 }
 
-static int publisher_main(int domainId, int sample_count)
+static int subscriber_main(int domainId, int sample_count)
 {
     DDS_DomainParticipant *participant = NULL;
-    DDS_Publisher *publisher = NULL;
+    DDS_Subscriber *subscriber = NULL;
     DDS_Topic *topic = NULL;
-    DDS_DataWriter *writer = NULL;
-    example2DataWriter *example2_writer = NULL;
-    example2 *instance = NULL;
+    struct DDS_DataReaderListener reader_listener =
+    DDS_DataReaderListener_INITIALIZER;
+    DDS_DataReader *reader = NULL;
     DDS_ReturnCode_t retcode;
-    DDS_InstanceHandle_t instance_handle = DDS_HANDLE_NIL;
     const char *type_name = NULL;
     int count = 0;
-    struct DDS_Duration_t send_period = {1, 0}; /* 1 s */
+    struct DDS_Duration_t poll_period = {1, 0}; /* 1 s */
     NDDS_Utility_NetworkCaptureParams_t params =
             NDDS_UTILITY_NETWORK_CAPTURE_PARAMETERS_DEFAULT;
 
@@ -107,15 +195,16 @@ static int publisher_main(int domainId, int sample_count)
 
     /*
      * Configuration of the capture.
-     *   - Only capture input traffic.
-     *   - Defaults: e.g. Do not parse encrypted content (i.e., show traffic
-     *     with protection, as in wire).
+     *   - Parse encrypted content.
+     *   - Remove all user data (serialized and encrypted) from the capture.
+     *   - Defaults: e.g., capture input and output traffic.
      */
-    params.traffic = NDDS_UTILITY_NETWORK_CAPTURE_TRAFFIC_IN;
-    if (!NDDS_Utility_start_network_capture_w_params("publisher", &params)) {
+    params.parse_encrypted_content = DDS_BOOLEAN_TRUE;
+    params.dropped_content =
+            NDDS_UTILITY_NETWORK_CAPTURE_CONTENT_ENCRYPTED_DATA;
+    if (!NDDS_Utility_start_network_capture_w_params("subscriber", &params)) {
         fprintf(stderr, "Error enabling network monitor for all participants\n");
     }
-
 
     /*
      * The example continues as the usual Connext DDS secure HelloWorld.
@@ -129,9 +218,9 @@ static int publisher_main(int domainId, int sample_count)
             DDS_TheParticipantFactory,
             &participant_qos,
             "SecurityExampleProfiles",
-            "B");
+            "A");
     if (retcode != DDS_RETCODE_OK) {
-        fprintf(stderr, "Unable to get participant qos from profile\n");
+        fprintf(stderr, "Unable to get default participant qos\n");
         return -1;
     }
     retcode = DDS_PropertyQosPolicyHelper_assert_pointer_property(
@@ -150,105 +239,85 @@ static int publisher_main(int domainId, int sample_count)
             NULL,
             DDS_STATUS_MASK_NONE);
     DDS_DomainParticipantQos_finalize(&participant_qos);
-
 #else
     participant = DDS_DomainParticipantFactory_create_participant_with_profile(
             DDS_TheParticipantFactory,
             domainId,
             "SecurityExampleProfiles",
-            "B",
+            "A",
             NULL,
             DDS_STATUS_MASK_NONE);
 #endif
     if (participant == NULL) {
         fprintf(stderr, "create_participant error\n");
-        publisher_shutdown(participant);
+        subscriber_shutdown(participant);
         return -1;
     }
 
-    publisher = DDS_DomainParticipant_create_publisher(
+    subscriber = DDS_DomainParticipant_create_subscriber(
             participant,
-            &DDS_PUBLISHER_QOS_DEFAULT,
+            &DDS_SUBSCRIBER_QOS_DEFAULT,
             NULL, /* listener */
             DDS_STATUS_MASK_NONE);
-    if (publisher == NULL) {
-        fprintf(stderr, "create_publisher error\n");
-        publisher_shutdown(participant);
+    if (subscriber == NULL) {
+        fprintf(stderr, "create_subscriber error\n");
+        subscriber_shutdown(participant);
         return -1;
     }
 
-    type_name = example2TypeSupport_get_type_name();
-    retcode = example2TypeSupport_register_type(
-            participant,
-            type_name);
+    type_name = NetworkCaptureTypeSupport_get_type_name();
+    retcode = NetworkCaptureTypeSupport_register_type(participant, type_name);
     if (retcode != DDS_RETCODE_OK) {
         fprintf(stderr, "register_type error %d\n", retcode);
-        publisher_shutdown(participant);
+        subscriber_shutdown(participant);
         return -1;
     }
 
     topic = DDS_DomainParticipant_create_topic(
             participant,
-            "Example example2",
+            "Example NetworkCapture using security",
             type_name,
             &DDS_TOPIC_QOS_DEFAULT,
             NULL, /* listener */
             DDS_STATUS_MASK_NONE);
     if (topic == NULL) {
         fprintf(stderr, "create_topic error\n");
-        publisher_shutdown(participant);
+        subscriber_shutdown(participant);
         return -1;
     }
 
-    writer = DDS_Publisher_create_datawriter(
-            publisher,
-            topic,
-            &DDS_DATAWRITER_QOS_DEFAULT,
-            NULL, /* listener */
-            DDS_STATUS_MASK_NONE);
-    if (writer == NULL) {
-        fprintf(stderr, "create_datawriter error\n");
-        publisher_shutdown(participant);
-        return -1;
-    }
-    example2_writer = example2DataWriter_narrow(writer);
-    if (example2_writer == NULL) {
-        fprintf(stderr, "DataWriter narrow error\n");
-        publisher_shutdown(participant);
-        return -1;
-    }
+    reader_listener.on_requested_deadline_missed  =
+    NetworkCaptureListener_on_requested_deadline_missed;
+    reader_listener.on_requested_incompatible_qos =
+    NetworkCaptureListener_on_requested_incompatible_qos;
+    reader_listener.on_sample_rejected =
+    NetworkCaptureListener_on_sample_rejected;
+    reader_listener.on_liveliness_changed =
+    NetworkCaptureListener_on_liveliness_changed;
+    reader_listener.on_sample_lost =
+    NetworkCaptureListener_on_sample_lost;
+    reader_listener.on_subscription_matched =
+    NetworkCaptureListener_on_subscription_matched;
+    reader_listener.on_data_available =
+    NetworkCaptureListener_on_data_available;
 
-    instance = example2TypeSupport_create_data_ex(DDS_BOOLEAN_TRUE);
-    if (instance == NULL) {
-        fprintf(stderr, "example2TypeSupport_create_data error\n");
-        publisher_shutdown(participant);
+    reader = DDS_Subscriber_create_datareader(
+            subscriber,
+            DDS_Topic_as_topicdescription(topic),
+            &DDS_DATAREADER_QOS_DEFAULT,
+            &reader_listener,
+            DDS_STATUS_MASK_ALL);
+    if (reader == NULL) {
+        fprintf(stderr, "create_datareader error\n");
+        subscriber_shutdown(participant);
         return -1;
     }
 
     for (count=0; (sample_count == 0) || (count < sample_count); ++count) {
+        printf("NetworkCapture subscriber sleeping for %d sec...\n",
+            poll_period.sec);
 
-        printf("Writing example2 Secure, count %d\n", count);
-
-        RTIOsapiUtility_snprintf(
-                instance->msg,
-                128,
-                "Hello World Secure (%d)",
-                count);
-
-        retcode = example2DataWriter_write(
-                example2_writer,
-                instance,
-                &instance_handle);
-        if (retcode != DDS_RETCODE_OK) {
-            fprintf(stderr, "write error %d\n", retcode);
-        }
-
-        NDDS_Utility_sleep(&send_period);
-    }
-
-    retcode = example2TypeSupport_delete_data_ex(instance, DDS_BOOLEAN_TRUE);
-    if (retcode != DDS_RETCODE_OK) {
-        fprintf(stderr, "example2TypeSupport_delete_data error %d\n", retcode);
+        NDDS_Utility_sleep(&poll_period);
     }
 
     /*
@@ -262,7 +331,7 @@ static int publisher_main(int domainId, int sample_count)
     /* Finalize the parameters to free allocated memory */
     NDDS_Utility_NetworkCaptureParams_t_finalize(&params);
 
-    return publisher_shutdown(participant);
+    return subscriber_shutdown(participant);
 }
 
 int main(int argc, char *argv[])
@@ -277,6 +346,6 @@ int main(int argc, char *argv[])
         sample_count = atoi(argv[2]);
     }
 
-    return publisher_main(domainId, sample_count);
+    return subscriber_main(domainId, sample_count);
 }
